+++ frontend/src/components/displayError.tsx
